requirements.txt
asgiref
Django
django-cors-headers
django-extensions
djangorestframework
gunicorn
packaging
pyt
sqlparse
typing_extensions
